backend/requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
